IMPORTANT CHANGES!!!
This is bug fix for removing students from enrolmentController. It fixes a major bug where removing a student only removed the student and related enrolments superficially, while they still remained in the editing context, were saved to the database and sent to the 4D database.

This bug probably only occured for queued payments (as the payment would have not seen the student and enrollment on the web site).

Anyway, now fixed and verified!