This is a preliminary commit of list view changes.
Things are still not fully working, but the Company view is the best implemented at the moment.

Things missing:
1) The buttons on the bottom of the list are not working.
2) Search, sort and delete are not working.

There is still quite a lot of work to be done... but I think I can limit this to the view controller classes (and list view associated classes such as ViewFrame, Panel, forms, etc) now.