This fixes the search dialog a little bit more... the mouse listener is being dealt with a little bit better, we are now using the view frame's default glass pane (as opposed to creating one). This incidentally fixed the weird widget/mouse pointer mismatch thing...
What it did involve is writing a bit of code that un/installs the default ViewFrame listeners... seems to work well... I need to rewrite the code that tries to block clicks on the ViewFrame, should be simple enough actually. There are also issues with the SearchQuery resize code... I'll also try to fix that.

It's nearly done anyway. I will use the more standardized approach in other dialogs.