Some more candy to the GUI.
OS X now plays nice with the toolbar.
The modal dialogs now play a beep when clicking on "disabled" bits of the frame.
The frame containing the modal dialog repaints when activated and also is now set to not resize when in modal mode.
Still need to fix WindowUtilities... not playing nice on OS X.