scripts update (index_templates_installer + projects) + updated script to point to new ISH index templates + added index template projects + added IndexTemplateRules.plist
It's now easy to keep Xcode Code Sense both aware and up-to-date with respect to the ISH frameworks
1) cd ~/Library/Application Support/Apple/Developer Tools/Index Templates/  && ln -s "~/Willow/frameworks/_Index Templates/IndexTemplateRules.plist

2) cd ~/Willow/scripts && ./index_templates_installer [-APPLE] -ISH

NOTE 1: if you've never installed the Apple index templates then pass the -APPLE argument which will also turn on the PBXCodeSenseEnableIndexTemplates default.

NOTE 2:
The first time this is run it takes a while - but subsequent runs are very quick because it only updates any changes. 

TIP:
I've got it running in a cron job every 2 minutes - and so never have to think about it again. No more complaints about out of date indexes :-)

I'm also utilising the build_project script to auto-build the frameworks at regular intervals. Easy.