WillowCollege & Frameworks + TRAVELEX Interaction.
NOTE 1: Interaction with the travelex gateway is functional and the records are correctly saved to the datastore. NEED TO TEST FURTHER.

NOTE 2: This update changes the model and associated methods so that primary & foreign keys are no longer exposed directly (the usual WO way).

Further to #2 other changes are in the works that will aid in the construction (and implied) qualifiers per entity.