Willow fixes for validateForSave.
It seems that to test for whether an object has truly changed you not only need to compare its currently snapshot to that of the previously stored snapshot but to also take into account that the snapshot dictionary contains changes for *relationships*.

Thus the isUpdatedRecord method in ISHGenericRecord now tests whether the set of keys made up from the object's attributes and toOne relationship keys intersects the set of keys from the snapshot diff dictionary - then we know that this object will change in the database. 

We're not interested in toMany changes because the foreign keys are not held in this object and thus won't change.