Adding near-complete PaymentManager to WillowSoap.
WillowEditingContextDelegate now uses a property to determine whether to observe notifications from all editing contexts EOEditingContextDidSaveChangesNotification or whether observations will be added/removed manually (as is needed for WillowSoap to limit queued records to those updated by the payment manager only).

The payment manager has everything in place (e.g., WOTimer to schedule itself for retrying payments), adding/removing itself as observer of notifications and so forth -- except for faulting in the payment/enrolments and actually processing the payment. This will be added once the api for payments has been adjusted with the addition of securepay to the mix.