All services now register themselves as observers of other services so that they can be notified when, for example, the application is going down - rather than polling the application continually.
This is especially helpful for when the application is quitting as all services are immediately notified that the application is going down and so they stop what they're doing immediately. Likewise, the replication services wait for a notification that the initial data population has completed before they do their thing.  Similarly, services dependant on derby being alive register for the notification of its status and so on.

The interface for being able to obtain the state when notifified is defined in ObservableState. Later on it will make more sense to have a central controller that receives all notifications and knows who needs to be told.