Populated Email Templates in General Preferences.
The first template is shown by default (currently) but for some reason it will not change templates when selecting another one from the popup (i.e., the event is fired the first time and then awaits the window to close before anything further happens!). So the event is getting trapped somewhere that I cannot figure as yet.

All templates are defined in MessageTemplate and available via the popup. There's also a selection of tags in the Available tags popup - but they do not do anything at this stage.

Marcin if you're able to have a look at this that'd be helpful. Thanks.