Adding BackManager to the live services.
The manager is kicked off along with the other services in the AngelServer constructor. At instantiation the manager registers itself as an observer of the application status and unregisters itself when exiting.

In practice this means that it will be notified when the application is going down and will internally cancel any scheduled tasks and notify the application that it has completed (or will notify the application when the current running tasks complete if any are running).

The ServerGUIController also registers itself as an observer of the BackupManager (like it does for the other services) so that it can display runtime info of a service's status.

The AngelServer when quitting awaits for all observers (besides DerbyServer) to complete their tasks (i.e., unregister themselves as app observers) before shutting down DerbyServer and quitting properly.

The BackupManager has two TODOs (both of which are in the resetSchedule method.
a) is simply about tying the iterator into the preferences so that the schedule is as preferred rather than hard-coded to 1am weekdays.
b) <annoymous new SchedulerTask()>.performBackup() must be done before anything will actually backup.

Otherwise it should be good to go. I've tested that starting and stopping the server successfully still functions - and it does.