--- Willow Payment to onCourse Invoice replication logic ---
Replication of payments from willow should now be possible. This will need to be thoroughly tested as it's critical that finances are correctly handled.

Invoice.updateFromSoap method on the server creates the Invoice, InvoiceLines, PaymentInLine, and PaymentIn objects from the incoming SOAPPayment and its associated enrolments. 

If any part causes an error the replication of that Payment fails - at least that's the intention. It's an error to replicate a payment twice (i.e., where the first replication succeeds).