Fixing ComboBox's loading of values and choices.
The order of loading is important in order to ensure that the stored data doesn't change unexpectedly by the comboBox's default behaviour of selecting the first item in the list (if no previous selection has been made).

In Controller:
1) Fire event to load stored value (if one exists). This (for example in the PreferenceFrameController) happens automatically via super.init() which in turn calls getValueForKey to obtain the desired value - and then fires the event to the comboBox.

2) Fire an event with the choices. The comboBox will automatically select the correct choice and _not_ save if the currentValue is not null. Otherwise it will select the first item in the popup and save.

If you reverse this order - you will be too late in retrieving the user's data from the db as it will have been replaced by a new selection.

We'll need to think a bit more about having the combo box automatically retrieve the choices upon receiving a value. For quick enrol - it would mean the controller notifying the component that new choices are available rather than providing a new value.