Adding support for upgrading onCourse data to be inline with the current version's validation rules.
Currently (because values are not pushed from server to client upon creation) virtually identical code is in client/server for onEntityCreation/onEntityModification methods.

Once we move to a version of Cayenne that supports pushing values from server to client - the client code should be removed (from the entities).

DerbyServer's Reference data population thread should iterate through the entities and for each entity test to see whether there are any objects that need upgrading (using a qualifier it obtains from the entity). If there are any objects these will be passed back to the entity for fixing and the thread will save the changes along the way.

This will need to be tested.