Switching PaymentInSummary's status combo for a non-editable textfield.
TextFieldWithLabel gains a new bean property: useChoicesKeyForDisplay.

To use this - the field must have this ticked and it must be uneditable. It will then use getChoicesForKey to figure out what to display for the given value (usually a constant).

Note: you'll need to override to return the default formatter for this component's property key in the controller.