ViewPanel now registers itself as a listener for CustomValueChangeEvents and gains a bean property: propertyKey.
PaymentInMainView listens for events for PaymentIn.TYPE_PROPERTY events. It then switches the inner payment panel(s) accordingly. 

ADDED BONUS: the controller doesn't even need to pass on the event to the view. The view receives the event when the user changes the selection in the type combo. So no additional code needed for this view. As such the code that was specifically calling this view (from both the QuickEnrollPaymentInController, PaymentInViewController, and PaymentInView) has been pulled.