Fixing some compile time errors for drag/drop events, adjusting case of static constants, & basic clean up (e.g., unused imports)
Drag/drop works - but the behaviour is not quite right for QE. e.g., Drag a student over QE a couple of times and the original data is lost. Also the events are quite right (or the Clairvoyance field isn't doing the right thing) as the related fields are only updating after the drop.

I'll look further into this whilst fixing the other clairvoyance issue we have to do with New... records.