Working around missing MissingSessionException [squish 5357].
See my comments here: http://issues.apache.org/cayenne/browse/CAY-671

Essentially, when the connection is lost, a CayenneRuntimeException is currently thrown - not the hoped (and new) MissingSessionException.

All is not lost, however, as the embedded cause [i.e., e.getCause()] is actually an instance of java.net.ConnectException.

So I've created a protected method in CayenneContext that each of the relevant methods utilise to figure out whether to throw on the runtime exception or do nothing. The method throwRuntimeException(Exception e) displays the connection lost dialog and returns false if e.getCause() instanceof ConnectException == true.