[squish 5176] adding SOAPPreference to the bean factories and adjusting replication algorithm to convert byte[] to NSData (as KeyValueCoding can sometimes not find the overloaded method).
The algorithm for both up and downstream replication now iterates through the public fields of the soap class - which should be faster than iterating through all the attributes of the underlying entity as we may not always replicate all fields. Likewise this allows for custom fields for soap that can be translated more easily.