[5726] Adding isDefault attribute to the Tax entity --> Updated schema wed version.
It is now no longer possible to change the code (or indeed the new boolean switch) for the default tax if the tax is the default tax. The preference is updated via post[Update | Persist] methods when necessary.

Some work also done for default Accounts - but this cannot be completed without a new screen in the preferences enabling the user to assign default accounts for various activities.