Fixing runtime exception thrown for JVM1.4.2 systems (calling MyClass.class.getSimpleName() is 1.5 and up only) + Consolidating report fetching and querying to a more common interface.
List views should be a little faster to open again now as we're only getting a count of reports when they open rather than fetching the reports regardless of whether they'll be used. OutputList will call the list controller's hasReports() function once in the constructor and retains that value so as to test whether the print button should be enabled ANDing that with whether there is a selection.