Restoring min size [squish 5367] behaviour & fixing frame closing bug [squish 5543] seems to be fixed now.
The getFrame().inModalState() is not quite correct for the purposes of the dispose method aborting in ViewController.

e.g., If I open an Account (Student Enrolment) and make an arbitrary change and press the window's close button - and press Discard changes on the sheet - then both the edit window would remain as well as the list window becoming visible again. This is  resolved by the following test.

hasFrame() && getFrame().hasPaletteLayer() || getFrame().hasModalLayer()

If ever that test succeeds then the edit window will remain open. Perhaps it should throw an exception so that events don't proceed as normal.