Adding mod_perl2 Auditorium Portal and removing the WebObjects equivalent.
This is a working copy of the mod_perl2 portal for streaming mp3 files. It includes a startup file (to preload modules at server launch), a handler for requests (ISH::Auditorium), and templates used by the module 'Template' (a front end to the Template Toolkit) for dynamic page generation.

The configuration file includes directives for WASM authentication, but they are not tested or complete.

The Apache2 LocationMatch is used to pass all requests (except for the /favicon.ico or /ish) to the script.