A. [squish 5844] B. [squish 5976]
A. Adding support for g[un]zipping soap request/responses. The technique could easily be used for other request handlers too as it has nothing to do with soap.

B. Adding sms message related entities and handlers to WillowSoap. 

Further work is needed on the [de]queueing mechanism so as to place responses back in the queue for replication back to angel.

NOTE: we will need to verify that angel will not (upon receiving objects from willow) put those same objects back into the queue again. Otherwise - we have the potential for an infinite loop.