Huge login optimisation (first impressions are important).
The PreferenceController no longer loads all user-related preferences at login. It instead loads/caches preferences as needed.

Comparison:
- old way: 25 seconds to login
- new way: 2 seconds tops. 1250% improvement!

Also, we're no longer deleting all old preferences and creating new ones when they have the same key. This  is a further optimisation that will help.

Now for the "BUT". This does mean that preferences saved from previous sessions that need to be loaded for the same view will now load via separate fetches rather than one big one at application load.

So, a future optimisation for such cases will be to have the preference controller load a set of preferences whose keys start with some substring. That'll speed things up at that point... but for now - at least the user gets into the app immediately so that their initial impression of the app is a good one.

We still have the problem of window positions and size not being restored properly - but that'll come later.