[squish 5976] adding preference for SMS delivery and so will not queue Message/MessagePerson if preference is false.
To accompany this, I've added to ReplicatableObject a [get|set]AllowsReplicationQueueing(...) method. These are implemented in CayenneDataObject and return Boolean.TRUE by default. 

For Message[Person] objects we override this to check the preference value if it's currently true.

The above "set" method has the additional benefit of allowing us to temporarily stop objects re-entering the queue when being received and updated from willow prior to addRecordToReplicationQueue being called.