[squish 5980] Server was not quitting correctly.
Quitting zero-conf at the beginning of the shutdown sequence was stalling the application. The HttpServer was not successfully stopping after that. 

I've now ensured that the HttpServer doesn't start unless AngelServer is in a START/STARTING state. I've moved the quit of zeroconf to after shutting down all other services (i.e., the end of AngelServer.quit).

In summary - AngelServer is now quitting properly again.

This would explain the problems we've been having with the scripts on delish and with control-c not always working.

Over to you Rex for further testing/verification.