[squish 5980] after failing to find the previous deadlock - this commit adds a new ServicesManager which acts sort of like LaunchServices on Mac OS X.
The idea is that each service has the following properties:
a. unique identifier (the full class name)
b. list of services that it depends on to run
c. list of services that it cannot run concurrently with

Otherwise a service is either a daemon or a periodic. Periodic services run their respecitive code (via the method runService()) every specified period. Daemon services are simply available all the time.

When shutting down, services shut down in the correct order - so that a service that has dependants (i.e., kids) will wait until the kids are down.

The main thread now waits until everything is down prior to exiting (which it was doing before - leaving zombie threads).

There'll be more optimisations in the future - such as getting threads to wait until notified rather than polling every so often for their status. But for now - it seems to work as expected. i.e., shuts down.

Some TODOs: This commit breaks the ServerGUIController's updating of the info text as we're starting up. So the rotary gif never turns off. I'll fix that now.