[squish 5980] onCourse server reenginered services. A ServicesManager keeps the state of and notifies other interested services of all services.
Each service itself is either a daemon service or a periodic service running one or more times periodically. Every service (or proxy service) sleeps until the state changes, in the case of daemon services, or for a specified interval or until the state changes (whichever comes first) for periodic services.

AuxiliaryServices are services that do not inherit from Service - and thus have a proxy service looking after it. Likewise the AuxiliaryService receives a proxy to communicate back with the service if necessary. e.g., AngelServer uses the proxy to notify its helper service that it's shutting down.

STARTING UP: (BRIEFLY)
- state_starting until dependent services up
- state_starting until excluded services down
- runServiceInitiasation(); // run pre-start init code
- notify of state
- daemon service: run until state stopping
  periodic service: run periodically while dependencies are up

STOPPING: e.g., via CMD+Q
- AngelServer [an auxiliary] notifies its proxy to stop and joins it's proxy thread; waiting until its dead.
- proxy service moves to state-stopping and notifies observers of its state
- each direct child service in turn moves to state-stopping and waits until children are dead
- when children are dead: runServiceFinalisation();
- move to state down and notify we're down
- ServerGUIController is the last to die. It's not connected to any services. i.e., has no dependencies.

In short - services wait no longer than the precise amount of time they have to wait for the state to change - making the shutdown and startup ultra fast.