Refactoring WillowCollege and related frameworks to handle uniquing per request rather than per application instance.
WillowWebObjects and WillowReusableComponents frameworks have been removed. Their files have been merged into WillowColleges for now. The future plan will be to rip out all components from WillowColleges and to have a framework for page components and a framework for reusable components. Overridden components (per site) are located in a separate directory.

A lot of cleanup can still occur; especially when we start to include ProjectWONDER frameworks in the mix.