[squish 6470] clairvoyance backspacing issues fixed. Null (e.g., for first name) is a valid value but was being ignored in the controller. This meant that when tabbing into another field and a loadData was called via an event on the previous field then the text that the user thought they'd deleted reappears.
[squish 6442] lostFocus is now doing the right thing. i.e., either (a) nothing if the focus event is temporary (e.g., window loses focus but when it regains focus we're back in business) or (b) updating the correct state according to the selection.

A number of other issues have been fixed as well as some bonus features being found along the way. These include:
a. code has had some further clean up including variables renamed into something more meaningful to save future confusion.

b. If we enter the state of isClairvoyanceEnabled() == false, then the textfield enters normal text editing mode. So backspaces and so forth behave normally (as they would in other text fields).

c. we no longer have the problem where entering some text into a field (e.g., course name) and then tabbing out would result in the text remaining in the field and being marked as an existing entity. i.e., if a record is not allowed to be created then we clear the field upon losing focus if the selected object is not an existing record.

d. if the list is not showing and we're not in New mode then pressing one of the normal list iterating keys (e.g., page down/up, up/down arrows or backspace) will make the list reappear as it should. This was previously only working for backspace characters. 

It would be possible for clairvoyance to automatically be switched back on using this same method if, for example, the user had typed a couple of characters in New mode and then changed their mind and pressed the down arrow to issue another check for an existing entity. That'll have to be a feature request...

Some further cleanup of the code will be necessary in the future to further readability and logic. This will involve double checking that common code is encapsulated in nicely named helper methods etc.