Adding a new property to ClairvoyanceField: "invokesRecordCreationHandler"
So, isAllowsRecordCreation() now controls whether a new data can be entered at all for the field.

and the new property determines whether we create a new controller to create the new record or just fire the event saying we've changed state.

In simpler terms - QE stuff will need to set the new property to false (the default is true). Everything else should then work as expected.