Fixing some logical bugs in Money and extending the roundWithScale method to allow for varying rounding types (than just round_half_up).
One bug related to [squish 6725] was not returning a Money for the round(Integer) method call for the case where it was calling the above method - leading to a ClassCastException elsewhere. The round(integer) method now returns a Money (even in signature) to guarentee this.

Some various other cleanups in Money.