Sharing common resource loading code (for client and server), enabling the version and wed versions to be available when running from classes (which will make it easier to test/simulate a real builds e.g., when profiling), and removing unused imports.
Note: I'm just adding the cayenne-src.jar so that we can easily navigate the source in your IDE of choice without having to individually build the thing.

Also fixing a race-condition when instantiating either AngelApp or AngelServer by setting the static variable for the app as the first thing in the constructor rather than in the main method.