Fixing exception thrown from AngelDequeuing where an object was in the wrong context when being deleted.
This should now fetch an object from a data row into the correct data context. 

The lifecycle changes have forced us to use ObjectContext as an interface - but there's a number of routines available from DataContext that are hidden :-/ So, for this I've had to rip out of a hidden cayenne class some functionality to support this operation.