Fixing the problem of drag-drop of a class onto the Quick Enrol screen.
The problem was that the transferred object was incomplete (any relation's ValueHolder was not set correctly because our data flavour is of type PersistentObject rather than the class of the object being transferred as far as I can tell).

I've adjusted all drop methods that don't already refault the object by getCustomComponentListener().getContext().localObject(dropObject.getObjectId()) before handing it on.

Also as a rule of thumb in our controllers we ought to do the following in valueChanged or setValueForKey:
don't do any operations on the PersistentObject prior to calling super. After calling super with the object, do a getValueForKey to retrieve the object that's been localised. That'll ensure that any drag-drop object has been refaulted correctly.

Essentially the ObjectId of an object is the only thing needed for dragging and dropping.

I've added getContext() to the CustomComponentListener interface.