fixed the race condition in initialising the list view.
The table was set to be visible with the valid reference to the increments list of records. The feature of that list is that at startup it is empty and the values in the list are fetched on request. Table.paint() run by EventDispatchThread was than faulting the objects, while the Foxtrot thread was doing the same causing Exceptions.

I have postponed the rendering of the table until the foxtrot thread is finished with faulting.

The way I did it is intercept the paint and in my further plans draw a infinite progress on top of empty table. Right now it just paints the empty table, the progress does not show up ... 

[Squish 6785]