Fixing Session.getID() for error 27 Mar 07 8:02:35 PM.
-----------------------
From: ari
Notes: double existing click event in room timetable view and then close it. 
------------------------

Note line 533 in ViewController where it's telling the ownerController to refresh components which is causing a valueChange event on the calendar component. i.e., getting it to reload its data. Why does it have a temp id if it's an existing record?