This should fix a session race condition (where the request context was unknown at the point of creating the WillowEditingContext).
Also tidying up/merging qualifier factories defined in both WillowEOAccess framework and the WillowColleges application.

There is also now a method Application.myApplication().isHTTPSEnabled(). This is utilised by some components to determine whether or not to display https hyperlinks and whether the page is required as secure. This can be optimised later by allowing this to be a preference per college (allowing for pre-deployment testing without having to assign another ip).

Side note: the ant/build.help.xml file now relies on a property plugins.home (which is the path to your eclipse plugins dir. Both plugins.home and eclipse.home should be added to your  ~/Library/wobuild.properties. plugins.dir may not be in your eclipse dir of course.