[squish 6807] fixing (seemingly) shutdown issue.
Re-reading the javadocs for Shutdown handlers was helpful. Essentially the shutdown handler should do its thing quickly. I'm pretty certain that this means it shouldn't wait until the main thread has exited. However, we still need an AWT-Event to join to the main thread so the if statement in handleQuit may change to that later.

Along the way, I've cleanup up console printing a little bit. To print a message on the console (i.e., for general System.out) call RuntimeUtil.println( object );

I've added a couple of helpful messages (like when the data population is finished and clients can connect so that you know where it's up to). These can be enhanced later.