Fixing inheritance model (via EOF work-a-round).
I'm 99% certain that there's a bug with EOF's vertical inheritance[1]. The good news is that I've been able to both narrow it down to an exact scenario and, thankfully, found an easy work-a-round [2]. Even better, we don't have to throw away the benefits for our tagging dreams.

This commit cleans up the inheritance slightly as well as providing for the work-a-round.

[1] See here for various links:
 http://issues.objectstyle.org/jira/browse/WOL-461
[2] simply requires exposing inherited/flattened foreign keys as class members allowing EOF to generate the correct sql on insert.