Committing mostly fixed support for SOAPTagRelation + cleaning up some weird model artifacts (some of the column keys were somehow wrong)
This adds support for bit-shifting compound primary keys into a single willowId for uniqueness. The maximum id possible in either primary key is (2 << 31) - 1. We've got a while.

The final piece in the puzzle is adding a hints dictionary to create the proper relationship to the taggable sub-entity. That'll be committed soon.