[squish 6807] server shutdown - zeroconf lockup.
Okay one of the common things I've seen from the thread dumps is that JmDNS was stuck during its call to close. One of the reasons for this could have been the static locks we had on our methods as it was waiting on a lock.

In either case, I've now implemented a non-lifetime-blocking closer for Zeroconf so that this doesn't block the rest of the shutdown but only waits a maximum of a few seconds before killing that shutdown thread.