[squish 7202] fixing jdk detection for both win and mac.
Essentially this boiled down to two things:
a) javac seems to print to stderr instead of stdout when printing the version. So we were previously getting no results on the input.
b) as javac also prints out help information (rather than just the version) I've trimmed the output down to just the bit with the version number for comparison.

This should work on both platforms now where the javac command is on the user's system path.