*** Willow Dynamic Colleges: first cut ***
This is a major rewrite of willow drastically simplifying design and deployment. As its name suggests the application provides a highly flexible system for building components without the need for a supporting java file in most cases (which would require a recompile).

I've provided the basic building blocks for building the app quickly. There will of course be a library of re-usable utility components bundled with the app (in frameworks) that will be useful within your dynamic components.

The documentation for DynamicComponents is in flux at the moment but will be fleshed out in the near future. The basic idea, however, is quite simple... there are 3 types of dynamic components: list, detail, and cookie-based. Each type can be embedded within another type as deep as you like.

In the end there is much less Java code than before (as I've gotten rid of, for the moment anyway, all the controllers which were inflexible)... more details later.