BinaryData replication handling.
For a BinaryData record (limited to one at a time currently - though that's easily extended) its content is replicated as an attachment (for encoding safety and all-round efficiency it seems) and reunited with its SOAPBinaryData object upon request before handing it off to the usual replication handling.

This requires the presence of the activation.jar and mail.jar in the classpath at both ends.