First-cut of EnrolmentMain component and accompanying logic.
ShortlistControl and CourseClassListItem components have been adjusted (but not tested) so as to handle the session actions when appropriate.

repetitions for students, classes, enrolments have been broken down into sub-components (for list items) to ease with ajax updates when they come - as well as sharing the layout with the existing components.