- fixed the glasspane progres showing, the size was 0x0 sometimes - more JUnit tests in regards to payment/enrolment processing/contacting willow.
The testing framework is now fully capable of any testing on client, except for gui testing, which means we can test creating records, deleting records. I think it will be very very *very* usable when we would add some tests which will verify triggers and all the financial data saving.

If our data is fine, than the gui is just a way to display this information, and than I believe the gui testing becomes less urgent.