merged the following changes from 0.7 branch to the trunk: - svn 4238: Commit by: lachlan
It seems that there are some queued records turning up that don't have a table_name column value. I've put in some safety for that - but still logging it as an error, and deleting the queued record.

That column should not allow null, so that the actual offending code is readily found.

-svn 4239, 4240:
Commit by marcin
 
fixed layouts and TextAreaWithLabel data saving and loading