Fixing comparator for soap entities.
Note: the comparator mustn't return 0 for keys that are different - which is what was happening for any entity that wasn't specifically ordered... thus every key that wasn't in the specifically ordered list returned 0 and was then treated the same giving us an empty fetch list. Now if the orderings for two objects are undefined we at least compare their string values to ensure that if they are different then they are treated as such in at least ascending order.