Safety for edge-case replication issue.
If, when converting to the real object from a queued record, we get a runtime exception or some cast exception - we'll log an error, and delete the queued record.

In case we happen to remove all of the current batch in this fashion then we silently move onto the next batch by resetting the current raw batch.