reviewing of svn 4362. Removed createObjectId from SOAPUtils, which was originally in ish.cayenne.access.ObjectContext
The story....

createObjectId was originally in ish.cayenne.access.ObjectContext prior to refactoring. It was utilised by objectForDataRow which was a method unavailable on an ObjectContext but existed in DataContext and so I'd added it to the ObjectContext interface. However since we had to wrap the context previously due to lifecycle events requirements implementing that method required access to the above method (which was in a private api of cayenne - so I'd copied it and shortened it for our purposes.

Now that lifecycle events is handled properly by DataContext there's no longer any need for us to keep a copy of the method.