[squish 7680]
Finished the work on multi-column sorting. It works on both data which is sorted in memory and that which is sorted in the db. Hold down shift to sort by a second column which keeping the first sorted. Pretty obscure, but other UI can come later - besides it isn't that important.

At the same table the table model classes have been documented a cleaned up considerably.