[squish 7621] enabling default sort orderings per shared entity [1] + fixing logic of cache invalidation
[1] subclassing EODatabaseContext.objectsWithFetchSpecification so that when a fault is fired and/or a sort orderings array is not supplied then if it's the shared ec then supplying the sort orderings array as defined from Properties. This is turned on/off via properties so it's safe.

This should mean that when following a toMany keypath (e.g., course.courseClasses) that the array returned from that keyPath will be sorted in the same way as if we'd fetched them manually via a fetchspecification with the sort orderings we prefer.

In essence this utilises database sorting. Note that as onCourse web scales up that this sorting will be better destributed to in-memory sorting via the various instances. But for now this will do. Another option could be utilising delegate methods for the faults to do the sorting in memory - but this was the easiest option for now.