yet another merge commit
the sessions count was displayed incorrectly, the getter method was failing to fetch the right value.
[Squish 7692]

duplicate preferences exception avoided.
[Squish 7687]

session count and minutes per session fields are disabled for classes with sessions defined. (utilised the implementation which was there for disabling the class start and end dates)

information message when cancelling or refunding of an enrolment fails improved