Implementation of Money.
OK. There is a lot in this one. Some notes:

* some problem I've not yet got to the bottom of is that I cannot rebuild the client Cayenne classes with cgen except by:

ant -lib ../common/classes/:../common/lib/

Some sort of classpath problem obviously, but cgen will just not find ish.math.Money as a custom class. Works on the server, not on the client. I suspect possibly a Cayenne problem but I've not yet investigated enough. Odd that identical code works on the server.

* Money no longer subclasses BigDecimal, but instead Number. This solved all Hessian issues in one simple act, but resulted in me touching lots of code.

* Issues remain:
  - list views do not display a formatted $xxx.xx but just raw numbers
  - edit views do not display the value of Money fields at all

I'm committing this since perhaps Marcin could help me nail some of the remaining issues.

On the plus side this opens the door to nice things like storing the currency type in the database.

Note that Money can never store a null value. I believe this is appropriate for currency calculations throughout onCourse. Creating it with null results in $0.