Adjusting WillowText for speed + timezone option + updating CourseClassDetail/ListItem to utilise WillowText where timezone info is needed.
In fact, it is possible to use WillowText wherever WOString could be used. So, I'll be changing all WOString usages to WillowText in the near future.

It's now possible to supply a timezone (i.e., its name) in conjunction with the timeformat. If no timezone is supplied then the default timezone is used. You'll find that for the entities that include timestamps (like start/endDates) that there's also a timeZone field. So passing this additional field allows us to present the time information appropriate for Perth, for example.

CourseClassEndDate: WillowText {
	escapeHTML = false;
	value = object.endDate;
	dateformat = "%a %e %b %y, %I:%M %p";
	timezone = object.timeZone;
}